refactor: move logic to shared-base
